requirements.txt
SQLAlchemy>=2.0.30
psycopg[binary,pool]>=3.2
alembic>=1.13.2
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
rich>=13.7.1
black==24.8.0